chore: 이미지 크기 변경 fix: 데스크탑 크기일 때와 아닐 때 커버 이미지가 따로 관리되는 문제 수정 	modified:   src/app/(contentWithoutScroll)/plans/[articleId]/_components/PlanDetailContent.tsx 	modified:   src/app/(contentWithoutScroll)/plans/[articleId]/_components/PlanDetailContentCoverImage.tsx 	modified:   src/app/(contentWithoutScroll)/plans/[articleId]/_components/PlanDetailContentHeader.tsx 	modified:   src/components/common/NextImage.tsx 	modified:   src/components/common/cards/PlanCard.tsx
